fix : benefit 타입 Long -> int 수정

+++ src/main/java/com/ureca/ocean/jjh/aibackend/client/dto/StoreUsageDto.java
@@ -18,5 +18,5 @@ public class StoreUsageDto {
     private String storeId;
     private UUID benefitId;
     private LocalDateTime visitedAt;
-    private Long benefit;
+    private int benefit;
 }
